Rename module to avoid conflicts
